IN-756 separate out cognito client
